+++ packages/backend/src/ee/services/ai/tools/findExplores.ts
@@ -64,8 +64,10 @@ const generateExploreResponse = ({
 `.trim()
             : ''
     }
-
-    <Fields alt="All fields from all tables" totalCount="${
+    ${
+        dimensions && metrics && dimensions.length > 0 && metrics.length > 0
+            ? `
+    <Fields alt="most popular dimensions and metrics" totalCount="${
         (dimensionsPagination?.totalResults ?? 0) +
         (metricsPagination?.totalResults ?? 0)
     }" displayedResults="${dimensions.length + metrics.length}">
@@ -100,23 +102,15 @@ const generateExploreResponse = ({
                 )
                 .join('\n            ')}
         </Metrics>
-    </Fields>
+    </Fields>`.trim()
+            : ''
+    }
 </Explore>`.trim();
 
-export const getFindExplores = ({
-    findExplores,
-    pageSize,
-    maxDescriptionLength,
-    fieldSearchSize,
-    fieldOverviewSearchSize,
-}: Dependencies) => {
-    const schema = toolFindExploresArgsSchema;
-
-    return tool({
-        description: `Tool: findExplores
+export const toolFindExploresDescription = `Tool: findExplores
 
 Purpose:
-Lists available Explores along with their field labels, joined tables, hints for you (Ai Hints), descriptions, and a sample set of dimensions and metrics.
+Lists available Explores along with their field labels, joined tables, hints for you (Ai Hints) and descriptions.
 
 Usage Tips:
 - Use this to understand the structure of an Explore before calling findFields.
@@ -124,8 +118,18 @@ Usage Tips:
 - Results are paginated — use the next page token to retrieve additional pages.
 - It's advised to look for tables first and then use the exploreName parameter to narrow results to a specific Explore.
 - When using the exploreName parameter, all fields and full description are returned for that explore.
-`,
-        parameters: schema,
+`;
+
+export const getFindExplores = ({
+    findExplores,
+    pageSize,
+    maxDescriptionLength,
+    fieldSearchSize,
+    fieldOverviewSearchSize,
+}: Dependencies) =>
+    tool({
+        description: toolFindExploresDescription,
+        parameters: toolFindExploresArgsSchema,
         execute: async (args) => {
             try {
                 if (args.page && args.page < 1) {
@@ -136,7 +140,7 @@ Usage Tips:
                     tableName: args.exploreName,
                     page: args.page ?? 1,
                     pageSize,
-                    includeFields: true,
+                    includeFields: !!args.exploreName,
                     fieldSearchSize,
                     fieldOverviewSearchSize,
                 });
@@ -165,4 +169,3 @@ ${exploreResponses}
             }
         },
     });
-};